Start database restructure
