+++ mdsuite/calculators/kirkwood_buff_integrals.py
@@ -14,7 +14,7 @@ import numpy as np
 import os
 import matplotlib.pyplot as plt
 from typing import Union
-from mdsuite.database.properties_database import PropertiesDatabase
+from mdsuite.database.property_database import PropertiesDatabase
 from mdsuite.database.database_scheme import SystemProperty
 from mdsuite.utils.exceptions import NotApplicableToAnalysis
 from mdsuite.calculators.calculator import Calculator
